+++ packages/sogrim-app/src/components/Banner/BannerCards/DegreeMainStatus.tsx
@@ -1,39 +1,19 @@
-import {
-  Button,
-  Card,
-  CardContent,
-  Typography,
-  Switch,
-  Tooltip,
-  Box,
-} from "@mui/material";
+import { Button, Card, CardContent, Typography, Box } from "@mui/material";
 import { DegreeStatusBar } from "./DegreeStatusBar";
 import { useState, useEffect } from "react";
 import { useStore } from "../../../hooks/useStore";
 import { observer } from "mobx-react-lite";
-import useUpdateUserSettings from "../../../hooks/apiHooks/useUpdateUserSettings";
-import { useAuth } from "../../../hooks/useAuth";
-import Confetti from "react-confetti";
+import { ComputeInProgressToggle } from "./ComputeInProgressToggle";
 
 const DegreeMainStatusComp: React.FC = () => {
   const {
-    dataStore: {
-      userDetails,
-      userSettings,
-      updateComputeInProgressInUserSettings,
-      isDegreeComplete,
-    },
+    dataStore: { userDetails, userSettings },
   } = useStore();
-
-  const { userAuthToken } = useAuth();
-  const { mutate, isError, error } = useUpdateUserSettings(userAuthToken);
 
   const [totalCredit, setTotalCredit] = useState<number>(0);
   const [onlyCompleteCredit, setOnlyCompleteCredit] = useState<number>(0);
   const [pointsDone, setPointsDone] = useState<number>(0);
   const [catalogName, setCatalogName] = useState<string>("");
-  const [confetti, setConfetti] = useState(false);
-  const [confettiRecycle, setConfettiRecycle] = useState(false);
 
   const [showMainStatus, setShowMainStatus] = useState<boolean>(false);
   const [computeInProgress, setComputeInProgress] = useState<boolean>(
@@ -62,23 +42,7 @@ const DegreeMainStatusComp: React.FC = () => {
       setCatalogName(catalogName);
       setShowMainStatus(catalogName !== "");
     }
-    if (isError) {
-      if ((error as any).response.status === 401) {
-        window.location.reload();
-      }
-    }
-
-    if (isDegreeComplete()) {
-      setConfetti(true);
-      setConfettiRecycle(true);
-      setTimeout(() => {
-        setConfettiRecycle(false);
-      }, 3000);
-    }
   }, [
-    error,
-    isDegreeComplete,
-    isError,
     pointsDone,
     userDetails,
     userDetails?.degree_status.course_statuses,
@@ -93,16 +57,6 @@ const DegreeMainStatusComp: React.FC = () => {
       ? 100
       : (onlyCompleteCredit / totalCredit) * 100;
 
-  const handleChange = (
-    _: React.SyntheticEvent,
-    computeInProgress: boolean
-  ) => {
-    setComputeInProgress(computeInProgress);
-    let newUserSettings =
-      updateComputeInProgressInUserSettings(computeInProgress);
-    mutate(newUserSettings);
-  };
-
   return showMainStatus ? (
     <Card sx={{ minWidth: 275, maxHeight: 150 }}>
       <CardContent>
@@ -110,29 +64,10 @@ const DegreeMainStatusComp: React.FC = () => {
           <Typography sx={{ fontSize: 18 }} color="text.secondary" gutterBottom>
             סטאטוס תואר
           </Typography>
-          <Tooltip
-            arrow
-            title={
-              "כפתור זה מאפשר לכם לחשב את סטטוס התואר שלכם עם התחשבות בקורסים שעוד אין להם ציון (בתהליך)"
-            }
-          >
-            <Box
-              sx={{
-                display: "flex",
-                flexDirection: "column",
-                alignItems: "center",
-                marginTop: "-10px",
-              }}
-            >
-              <Typography fontSize={10}>התחשב בקורסים בתהליך</Typography>
-              <Switch
-                onChange={handleChange}
-                checked={computeInProgress}
-                color="secondary"
-                size="small"
-              />
-            </Box>
-          </Tooltip>
+          <ComputeInProgressToggle
+            computeInProgress={computeInProgress}
+            setComputeInProgress={setComputeInProgress}
+          />
         </Box>
         <DegreeStatusBar
           {...{
@@ -147,14 +82,6 @@ const DegreeMainStatusComp: React.FC = () => {
         <Button sx={{ display: "flex", justifyContent: "center" }} size="small">
           {catalogName}
         </Button>
-        {confetti && (
-          <Confetti
-            width={2000}
-            numberOfPieces={500}
-            recycle={confettiRecycle}
-            onConfettiComplete={() => setConfetti(false)}
-          />
-        )}
       </CardContent>
     </Card>
   ) : null;
